// frontend/src/client/models/ObservationOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ObservationOut = {
    location: string;
    latitude: number;
    longitude: number;
    date_time: string;
    pH: number;
    conductivity: number;
    description: string;
    DO: number;
    contaminants: string;
    id: number;
    owner_id: number;
};