// frontend/src/client/models/ObservationUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ObservationUpdate = {
    location: string;
    latitude: number;
    longitude: number;
    date_time: string;
    pH: number;
    conductivity: number;
    description: string;
    DO: number;
    contaminants: string;
};